Early stages of adding ball + spawner
Added a spawner that spawns a ball when a key is pressed. The ball is destroyed if they suck at hitting ( so they can try again). Also made it so that it isn't destroyed if they actaully hit it ( Currently you hit by pressing a key (nothing happens)).
There is a physics material on ball and terrain